requirements.txt
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
whitenoise==5.0.1
gunicorn==20.0.4
PyJWT==1.7.1
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
django-allauth==0.42.0